specular lights

+++ src/light.hpp
@@ -11,7 +11,7 @@ public:
     : m_color{glm::vec3(red, green, blue)}, m_ambient_intensity{intensity}, m_direction{glm::vec3(x_dir, y_dir, z_dir)}, m_diffuse_intensity(diffuse_intensity)
     {}
 
-    void use_light(GLfloat ambient_intensity_location, GLfloat ambient_color_location, GLfloat diffuse_intensity_location, GLfloat direction_location)
+    void use_light(GLuint ambient_intensity_location, GLuint ambient_color_location, GLuint diffuse_intensity_location, GLuint direction_location)
     {
         glUniform3f(ambient_color_location, m_color.x, m_color.y, m_color.z);
         glUniform1f(ambient_intensity_location, m_ambient_intensity);
